w3c-reports/12092014/mac-wk2/runned-tests/2dcontext/compositing/2d.composite.image.destination-out-expected.txt

2d.composite.image.destination-out
Actual output:
Expected output:


FAIL Canvas test: 2d.composite.image.destination-out assert_approx_equals: Alpha channel of the pixel at (50, 25) expected 32 +/- 5 but got 127(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_approx_equals@http://localhost:8800/resources/testharness.js:425:15
_assertPixelApprox@http://localhost:8800/common/canvas-tests.js:50:25
http://localhost:8800/2dcontext/compositing/2d.composite.image.destination-out.html:27:19
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/common/canvas-tests.js:62:19
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/common/canvas-tests.js:59:15)
